# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LivroLog API",
        "description": "Personal library management system with Google Books API integration",
        "contact": {
            "email": "support@livrolog.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "LivroLog API Server"
        }
    ],
    "paths": {
...
        "/api/authors/merge": {
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Merge two duplicate authors into one",
                "description": "Moves all books from the source author to the target author and deletes the duplicate author. **Admin access required.**",
                "operationId": "d25e7ab98ccb891844da86916cce44b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "from_author_id",
                                    "to_author_id"
                                ],
                                "properties": {
                                    "from_author_id": {
                                        "description": "ID of the duplicate author to be merged (will be deleted)",
                                        "type": "string",
                                        "example": "A-9IO8-3D6Y"
                                    },
                                    "to_author_id": {
                                        "description": "ID of the main author to keep",
                                        "type": "string",
                                        "example": "A-1ABC-2DEF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authors merged successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Authors merged successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Access denied. Admin privileges required."
                    },
                    "422": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insert the access token obtained from login endpoint. Format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}